+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Set default template to `gem_layout` and remove references to the deprecated `core_layout`.
+
 # 16.0.1
 * Resolve Nokogiri deprecation warning (#277)